+++ refinedstorage-common-api/src/main/java/com/refinedmods/refinedstorage/common/api/RefinedStorageApi.java
@@ -208,6 +208,8 @@ public interface RefinedStorageApi {
 
     ResourceLocation getCreativeModeTabId();
 
+    ResourceLocation getColoredCreativeModeTabId();
+
     AbstractNetworkNodeContainerBlockEntity<?> createStorageBlockEntity(BlockPos pos,
                                                                         BlockState state,
                                                                         StorageBlockProvider provider);